+++ packages/editor/api-report.api.md
@@ -104,9 +104,6 @@ import { whyAmIRunning } from '@tldraw/state';
 // @internal (undocumented)
 export function activeElementShouldCaptureKeys(): boolean;
 
-// @public (undocumented)
-export type AdjacentDirection = 'down' | 'left' | 'next' | 'prev' | 'right' | 'up';
-
 // @public
 export function angleDistance(fromAngle: number, toAngle: number, direction: number): number;
 
@@ -1242,6 +1239,7 @@ export class Editor extends EventEmitter<TLEventMap> {
     // @internal (undocumented)
     getCurrentPageShapeIdsSorted(): TLShapeId[];
     getCurrentPageShapes(): TLShape[];
+    getCurrentPageShapesInReadingOrder(): TLShape[];
     getCurrentPageShapesSorted(): TLShape[];
     getCurrentPageState(): TLInstancePageState;
     getCurrentTool(): StateNode;
@@ -1270,6 +1268,7 @@ export class Editor extends EventEmitter<TLEventMap> {
     getIsReadonly(): boolean;
     // @internal
     getMarkIdMatching(idSubstring: string): null | string;
+    getNearestAdjacentShape(currentShapeId: TLShapeId, direction: 'down' | 'left' | 'right' | 'up'): TLShapeId;
     getOnlySelectedShape(): null | TLShape;
     getOnlySelectedShapeId(): null | TLShapeId;
     // @deprecated (undocumented)
@@ -1480,6 +1479,8 @@ export class Editor extends EventEmitter<TLEventMap> {
     screenToPage(point: VecLike): Vec;
     readonly scribbles: ScribbleManager;
     select(...shapes: TLShape[] | TLShapeId[]): this;
+    // (undocumented)
+    selectAdjacentShape(direction: TLAdjacentDirection): void;
     selectAll(): this;
     selectNone(): this;
     sendBackward(shapes: TLShape[] | TLShapeId[], opts?: {
@@ -1587,7 +1588,7 @@ export class Editor extends EventEmitter<TLEventMap> {
     } & TLCameraMoveOptions): this;
     zoomToFit(opts?: TLCameraMoveOptions): this;
     zoomToSelection(opts?: TLCameraMoveOptions): this;
-    zoomToShapeIfOffscreen(padding?: number): void;
+    zoomToSelectionIfOffscreen(padding?: number): void;
     zoomToUser(userId: string, opts?: TLCameraMoveOptions): this;
 }
 
@@ -1778,9 +1779,6 @@ export function getFreshUserPreferences(): TLUserPreferences;
 // @public
 export function getIncrementedName(name: string, others: string[]): string;
 
-// @public
-export function getNearestAdjacentShape(editor: Editor, currentShapeId: TLShapeId, direction: 'down' | 'left' | 'right' | 'up'): TLShapeId;
-
 // @public (undocumented)
 export function getPerfectDashProps(totalLength: number, strokeWidth: number, opts?: Partial<{
     closed: boolean;
@@ -1829,9 +1827,6 @@ export function getRotationSnapshot({ editor, ids, }: {
     editor: Editor;
     ids: TLShapeId[];
 }): null | TLRotationSnapshot;
-
-// @public
-export function getShapesInReadingOrder(editor: Editor): TLShape[];
 
 // @public (undocumented)
 export function getSnapshot(store: TLStore): TLEditorSnapshot;
@@ -2546,9 +2541,6 @@ export class ScribbleManager {
 }
 
 // @public (undocumented)
-export function selectAdjacentShape(editor: Editor, direction: AdjacentDirection): void;
-
-// @public (undocumented)
 export type SelectionCorner = 'bottom_left' | 'bottom_right' | 'top_left' | 'top_right';
 
 // @public (undocumented)
@@ -2925,6 +2917,9 @@ export type TiptapEditor = Editor_2;
 
 // @public
 export type TiptapNode = Node_2;
+
+// @public (undocumented)
+export type TLAdjacentDirection = 'down' | 'left' | 'next' | 'prev' | 'right' | 'up';
 
 // @public (undocumented)
 export type TLAnyBindingUtilConstructor = TLBindingUtilConstructor<any>;
