+++ packages/tldraw/src/lib/ui/context/actions.tsx
@@ -17,7 +17,6 @@ import {
 	compact,
 	createShapeId,
 	openWindow,
-	selectAdjacentShape,
 	useMaybeEditor,
 	useValue,
 } from '@tldraw/editor'
@@ -1486,7 +1485,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: '$left',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'left')
+						editor.selectAdjacentShape('left')
 						trackEvent('select-adjacent-shape', { source, direction: 'left' })
 					}
 				},
@@ -1496,7 +1495,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: '$right',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'right')
+						editor.selectAdjacentShape('right')
 						trackEvent('select-adjacent-shape', { source, direction: 'right' })
 					}
 				},
@@ -1506,7 +1505,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: '$up',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'up')
+						editor.selectAdjacentShape('up')
 						trackEvent('select-adjacent-shape', { source, direction: 'up' })
 					}
 				},
@@ -1516,7 +1515,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: '$down',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'down')
+						editor.selectAdjacentShape('down')
 						trackEvent('select-adjacent-shape', { source, direction: 'down' })
 					}
 				},
@@ -1526,7 +1525,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: 'tab',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'next')
+						editor.selectAdjacentShape('next')
 						trackEvent('select-adjacent-shape', { source, direction: 'next' })
 					}
 				},
@@ -1536,7 +1535,7 @@ export function ActionsProvider({ overrides, children }: ActionsProviderProps) {
 				kbd: '!tab',
 				onSelect(source) {
 					if (path.startsWith('select.')) {
-						selectAdjacentShape(editor, 'prev')
+						editor.selectAdjacentShape('prev')
 						trackEvent('select-adjacent-shape', { source, direction: 'prev' })
 					}
 				},
